# backend/docs/swagger.yaml
basePath: /
definitions:
  main.JSONResponse:
    properties:
      data:
        description: any is an alias for interface{}
      error:
        type: boolean
      message:
        type: string
    required:
    - error
    - message
    type: object
  models.UserCredentials:
    properties:
      email:
        example: admin@example.com
        type: string
      password:
        description: internally is stored as hash
        example: password
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8888
info:
  contact: {}
  description: Service for keeping track of your training progress.
  title: Workout Tracker
  version: 0.0.0
paths:
  /authenticate:
    post:
      consumes:
      - application/json
      description: Authenticate user by email address and password.
      parameters:
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UserCredentials'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/main.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.JSONResponse'
      summary: authenticate user
      tags:
      - auth
  /logout:
    get:
      description: Logout and remove refresh token from cookie storage.
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/main.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: logout
      tags:
      - auth
  /refresh:
    get:
      description: Refresh JWT token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.JSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.JSONResponse'
      summary: refresh jwt token
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"